+++ packages/amplify-util-uibuilder/src/commands/generateComponents.ts
@@ -16,6 +16,8 @@ import {
   isStudioForm,
   isFormDetachedFromModel,
   deleteDetachedForms,
+  hasStorageField,
+  isFormSchema,
 } from './utils';
 
 /**
@@ -60,6 +62,7 @@ export const run = async (context: $TSContext, eventType: 'PostPush' | 'PostPull
     let hasSuccessfulForm = false;
     const failedResponseNames: string[] = [];
     const modelNames = dataSchema?.models ? new Set(Object.keys(dataSchema.models)) : new Set<string>();
+    let hasStorageManagerField = false;
 
     Object.entries(generatedResults).forEach(([key, results]) => {
       results.forEach((result) => {
@@ -67,6 +70,10 @@ export const run = async (context: $TSContext, eventType: 'PostPush' | 'PostPull
           successfulSchemas.push(result.schema);
           if (key === 'form') {
             hasSuccessfulForm = true;
+
+            if (!hasStorageManagerField && isFormSchema(result.schema) && hasStorageField(result.schema)) {
+              hasStorageManagerField = true;
+            }
           }
         } else {
           const failedSchema = result.schema;
@@ -112,7 +119,7 @@ export const run = async (context: $TSContext, eventType: 'PostPush' | 'PostPull
       );
     }
 
-    notifyMissingPackages(context);
+    notifyMissingPackages(context, hasStorageManagerField);
 
     await deleteDetachedForms(detachedForms, studioClient);
   } catch (e) {